+++ django/template/defaulttags.py
@@ -486,7 +486,7 @@ class WidthRatioNode(Node):
             value = float(value)
             max_value = float(max_value)
             ratio = (value / max_value) * max_width
-            result = str(int(round(ratio)))
+            result = str(round(ratio))
         except ZeroDivisionError:
             return '0'
         except (ValueError, TypeError, OverflowError):
